zlib1g
Previously, Java applications could not run on the tiny run image.  With the
advent of the GraalVM[1] Native Image[2] functionality, this is now a
possibility.  Native Images have few, but some library dependencies.  Most of
these libraries are already on the tiny run image, and two of the three that
are not (libstdc++ and libgcc_s) can be statically compiled into the native
image binary.  However, there is a single library dependency on libz that
cannot be compiled into the binary.

This change adds the zlib1g package to the base images and the tiny run image.
There are good reasons to minimize the collection of packages on base and
especially tiny, but given that this library is quite commonly used and is the
only barrier to Java native images on tiny I think it's a trade-off worth
making.  In addition, the list of libraries required by GraalVM native images
has been constant for some time, and while it's impossible to predict the
future, it is currently unlikely that any further libraries would need to be
added to tiny.

[1]: https://www.graalvm.org
[2]: https://www.graalvm.org/docs/reference-manual/native-image/

Signed-off-by: Ben Hale <bhale@vmware.com>